# src/configs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "habithub-api",
    "description": "habithub-express-mongoose",
    "termsOfService": "https://www.termsfeed.com/live/ed8b4e15-b05a-41d6-b12b-920a89756f29",
    "contact": {
      "name": "omerrfarukcapur@gmail.com",
      "email": "omerrfarukcapur@gmail.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "host": "127.0.0.1:8000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Token": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Simple Token Authentication * Example: <b>Token ...tokenKey...</b>"
    },
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Authentication * Example: <b>Bearer ...accessToken...</b>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "description": "\r  You can send query with endpoint for search[], sort[], page and limit.\r  <ul> Examples:\r  <li>URL/?<b>search[field1]=value1&search[field2]=value2</b></li>\r  <li>URL/?<b>sort[field1]=1&sort[field2]=-1</b></li>\r  <li>URL/?<b>page=2&limit=1</b></li>\r  </ul>\r  ",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Handles user feedback submission by validating input and sending feedback via email.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Name of the user"
                        },
                        "example": {
                          "type": "string",
                          "example": "John Doe"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "email"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email address of the user"
                        },
                        "example": {
                          "type": "string",
                          "example": "john.doe@example.com"
                        }
                      }
                    },
                    "subject": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Subject of the feedback"
                        },
                        "example": {
                          "type": "string",
                          "example": "Feedback Subject"
                        }
                      }
                    },
                    "feedback": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Feedback message"
                        },
                        "example": {
                          "type": "string",
                          "example": "This is a feedback message."
                        }
                      }
                    }
                  }
                },
                "required": {
                  "type": "array",
                  "example": [
                    "name",
                    "email",
                    "feedback"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Feedback submission data"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string",
            "example": "60c72b2f9f1b2c6f4c8e9b6a"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Single User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "description": "Handles user feedback submission by validating input and sending feedback via email.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Name of the user"
                        },
                        "example": {
                          "type": "string",
                          "example": "John Doe"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "email"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email address of the user"
                        },
                        "example": {
                          "type": "string",
                          "example": "john.doe@example.com"
                        }
                      }
                    },
                    "subject": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Subject of the feedback"
                        },
                        "example": {
                          "type": "string",
                          "example": "Feedback Subject"
                        }
                      }
                    },
                    "feedback": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Feedback message"
                        },
                        "example": {
                          "type": "string",
                          "example": "This is a feedback message."
                        }
                      }
                    }
                  }
                },
                "required": {
                  "type": "array",
                  "example": [
                    "name",
                    "email",
                    "feedback"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Feedback submission data"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string",
            "example": "60c72b2f9f1b2c6f4c8e9b6a"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "description": "Handles user feedback submission by validating input and sending feedback via email.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Name of the user"
                        },
                        "example": {
                          "type": "string",
                          "example": "John Doe"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "email"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email address of the user"
                        },
                        "example": {
                          "type": "string",
                          "example": "john.doe@example.com"
                        }
                      }
                    },
                    "subject": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Subject of the feedback"
                        },
                        "example": {
                          "type": "string",
                          "example": "Feedback Subject"
                        }
                      }
                    },
                    "feedback": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Feedback message"
                        },
                        "example": {
                          "type": "string",
                          "example": "This is a feedback message."
                        }
                      }
                    }
                  }
                },
                "required": {
                  "type": "array",
                  "example": [
                    "name",
                    "email",
                    "feedback"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Feedback submission data"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string",
            "example": "60c72b2f9f1b2c6f4c8e9b6a"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
...
  "security": [
    {
      "Token": []
    },
    {
      "Bearer": []
    }
  ]
}